platform/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pyjwt>=2.0.0
bcrypt>=4.0.0
